requirements.txt
h5py==3.7.0
imageio==2.21.1
keras==2.9.0
matplotlib==3.5.2
numpy==1.21.6
opencv-python==4.6.0.66
Pillow==9.1.0
PyWavelets==1.3.0
PyYAML==6.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sklearn==0.0
tensorflow==2.9.1
tensorflow-addons==0.17.1
tf-slim==1.1.0
tflearn==0.5.0
torch==1.12.1
tqdm==4.64.0

